requirements.txt
# ============================================
# Fractal Connector - Dependencies
# ============================================

# === CORE ===
websockets>=12.0
aiohttp>=3.9.0
flask>=3.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0
cryptography>=41.0.0
watchdog>=3.0.0
pandas>=2.1.0
requests>=2.31.0
psutil>=5.9.0

# === SQL DATABASES ===
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0      # PostgreSQL
pymysql>=1.1.0              # MySQL
pyodbc>=5.0.0               # SQL Server (ODBC)

# === NOSQL DATABASES ===
pymongo>=4.6.0              # MongoDB
redis>=5.0.0                # Redis
elasticsearch>=8.0.0        # Elasticsearch

# === CLOUD DATA WAREHOUSES ===
snowflake-connector-python>=3.5.0

# === CLOUD STORAGE ===
boto3>=1.34.0               # AWS S3
google-cloud-storage>=2.14.0  # Google Cloud Storage
azure-storage-blob>=12.19.0   # Azure Blob Storage

# === PRODUCTIVITY ===
google-api-python-client>=2.100.0  # Google Sheets
google-auth>=2.25.0

# === MESSAGING ===
kafka-python>=2.0.2         # Apache Kafka

# === FILE TRANSFER ===
paramiko>=3.4.0             # SFTP

# === MARKET DATA (Free) ===
yfinance>=0.2.0             # Yahoo Finance

# === TRADING (Optional - install as needed) ===
# blpapi>=3.0.0             # Bloomberg - requires Bloomberg SDK
# eikon>=1.1.0              # Refinitiv - requires Eikon license
# ib_insync>=0.9.0          # Interactive Brokers
# oracledb>=2.0.0           # Oracle Database

# === BUILD ===
pyinstaller>=6.0.0
